Comp fix + default window size relative to monitor

+++ voxel_engine/shaders/PixelShader.glsl
@@ -6,7 +6,7 @@ in vec3 pos;
 in vec3 normal;
 in vec2 uv;
 flat in uint texID;
-in int ao;
+in float ao;
 
 uniform sampler2DArray texArray;
 uniform float time;
